Andrewtlu/redirects (#15)
* Update login success redirect, update caniuse

* Move code verification out of page component for reset password

* Update createClient

* Extract signout to endpoint

* Redirect non-auth paths to login if not logged in

* Refactor login to be signin (semantics)

* Add comments to middleware

* Fix navbar dashboard link, redirect signin to dashboard if already authenticated

* Change 127.0.0.1 to localhost, clean login layout

* Update middleware auth check to be more lenient

* Remove signin route, update action message
