+++ .env.sample
@@ -1,4 +1,4 @@
 NEXT_PUBLIC_SUPABASE_URL=[INSERT SUPABASE PROJECT URL]
 NEXT_PUBLIC_SUPABASE_ANON_KEY=[INSERT SUPABASE PROJECT API ANON KEY]
 DATABASE_URL=[INSERT SUPABASE POSTGRES URL "psql://..."]
-NEXT_PUBLIC_SITE_URL=http://127.0.0.1:3000
+NEXT_PUBLIC_SITE_URL=http://localhost:3000
