--- src/app/(login)/layout.tsx
@@ -1,3 +1,0 @@
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
