+++ src/app/(signin)/forgot-password/action.ts
@@ -1,9 +1,9 @@
+"use server";
+
 import { createClient } from "@/utils/supabase/server";
 import { redirect } from "next/navigation";
 
 const resetPassword = async (formData: FormData) => {
-  "use server";
-
   const email = formData.get("email") as string;
   const supabase = createClient();
   const { error } = await supabase.auth.resetPasswordForEmail(email, {
