+++ src/app/(signin)/layout.tsx
@@ -1,0 +1,15 @@
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex w-full flex-1 flex-col justify-center gap-2 bg-maroon px-8">
+      <div className="mb-10">
+        <h1 className="pb-5 text-center font-playfair text-7xl text-white">
+          Odyssey Family Counseling
+        </h1>
+        <h2 className="text-center font-playfair text-5xl text-odyssey-yellow">
+          Employee Portal
+        </h2>
+      </div>
+      {children}
+    </div>
+  );
+}
