+++ src/app/(signin)/signin/action.ts
@@ -1,12 +1,14 @@
 "use server";
 
 import { createClient } from "@/utils/supabase/server";
+import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
 const signIn = async (formData: FormData) => {
+  const supabase = createClient();
+
   const email = formData.get("email") as string;
   const password = formData.get("password") as string;
-  const supabase = createClient();
 
   const { error } = await supabase.auth.signInWithPassword({
     email,
@@ -14,10 +16,11 @@ const signIn = async (formData: FormData) => {
   });
 
   if (error) {
-    return redirect("/login?message=Could not authenticate user");
+    return redirect(`/signin?message=${error.message}`);
   }
 
-  return redirect("/protected");
+  revalidatePath("/", "layout");
+  return redirect("/dashboard");
 };
 
 export default signIn;
