+++ src/app/auth/signout/route.ts
@@ -1,0 +1,19 @@
+import { createClient, useUser } from "@/utils/supabase/server";
+import { revalidatePath } from "next/cache";
+import { type NextRequest, NextResponse } from "next/server";
+
+export async function POST(req: NextRequest) {
+  const supabase = createClient();
+
+  // Check if a user's logged in
+  const user = await useUser();
+
+  if (user) {
+    await supabase.auth.signOut();
+  }
+
+  revalidatePath("/", "layout");
+  return NextResponse.redirect(new URL("/signin", req.url), {
+    status: 302,
+  });
+}
