+++ src/app/dashboard/page.tsx
@@ -1,21 +1,14 @@
 import { SubNavBar } from "@/components/NavBar";
-import { createClient } from "@/utils/supabase/server";
-import { redirect } from "next/navigation";
+import { useUser } from "@/utils/supabase/server";
 
 export default async function Index() {
-  const supabase = createClient();
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) {
-    redirect("/login");
-  }
+  const user = await useUser();
 
   return (
     <div className="flex w-full flex-1 flex-col items-center gap-20">
       <div className="animate-in flex max-w-4xl flex-1 flex-col gap-20 px-3 opacity-0">
-        <SubNavBar selected={`HR Services for ${user.email}`} />
+        {/* non-null assertion because middleware will redirect to signin if user doesn't exist */}
+        <SubNavBar selected={`HR Services for ${user!.email}`} />
       </div>
     </div>
   );
