+++ src/utils/supabase/server.ts
@@ -1,50 +1,34 @@
-import { createServerClient, type CookieOptions } from "@supabase/ssr";
-import { revalidatePath } from "next/cache";
-import { type RequestCookie } from "next/dist/compiled/@edge-runtime/cookies";
+import { createServerClient } from "@supabase/ssr";
 import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
 
-export const createClient = () => {
+export function createClient() {
   const cookieStore = cookies();
 
+  // Create a server's supabase client with newly configured cookie,
+  // which could be used to maintain user's session
   return createServerClient(
     process.env.NEXT_PUBLIC_SUPABASE_URL!,
     process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
     {
       cookies: {
-        get(name: string) {
-          return cookieStore.get(name)?.value;
+        getAll() {
+          return cookieStore.getAll()
         },
-        set(name: string, value: string, options: CookieOptions) {
+        setAll(cookiesToSet) {
           try {
-            cookieStore.set({ name, value, ...options } as RequestCookie);
-          } catch (error) {
-            // The `set` method was called from a Server Component.
-            // This can be ignored if you have middleware refreshing
-            // user sessions.
-          }
-        },
-        remove(name: string, options: CookieOptions) {
-          try {
-            cookieStore.set({ name, value: "", ...options } as RequestCookie);
-          } catch (error) {
-            // The `delete` method was called from a Server Component.
+            cookiesToSet.forEach(({ name, value, options }) =>
+              cookieStore.set(name, value, options)
+            )
+          } catch {
+            // The `setAll` method was called from a Server Component.
             // This can be ignored if you have middleware refreshing
             // user sessions.
           }
         },
       },
-    },
-  );
-};
-
-export const signOut = async () => {
-  const supabase = createClient();
-  await supabase.auth.signOut();
-
-  revalidatePath("/");
-  return redirect("/login");
-};
+    }
+  )
+}
 
 export const useUser = async () => {
   const supabase = createClient();
